Fixed viewport size and jump.

+++ core/src/com/my/game/MyGame.java
@@ -11,8 +11,8 @@ import com.my.game.screens.PlayScreen;
 public class MyGame extends Game {
 	//
 	public SpriteBatch batch;
-	public static final int V_WIDTH = 400;
-	public static final int V_HEIGHT = 208;
+	public static final float V_WIDTH = 400;
+	public static final float V_HEIGHT = 208;
 	public static final float PPM = 100;
 
 	/**
